third_party/WebKit/LayoutTests/fast/dom/HTMLElement/attr-missing-ancestor-false-expected.txt

When contentEditable is missing, it maps to "inherit" state and the element is NOT editable if its nearest ancestor is NOT.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById("p").hasAttribute("contentEditable") is false
PASS document.getElementById("p").contentEditable is "inherit"
PASS document.getElementById("p").isContentEditable is false
PASS window.getComputedStyle(p, "").getPropertyValue("-webkit-user-modify") is "read-only"
PASS successfullyParsed is true

TEST COMPLETE
